+++ t/app/controller/admin/bodies.t
@@ -232,6 +232,23 @@ FixMyStreet::override_config {
         is $contact->get_extra_metadata('group'), 'group a', "group stored correctly";
     };
 
+    subtest 'group can be unset' => sub {
+        $mech->get_ok('/admin/body/' . $body->id);
+        $mech->content_contains( 'group</strong> is used for the top-level category' );
+
+        $mech->submit_form_ok( { with_fields => {
+            category   => 'grouped category',
+            email      => 'test@example.com',
+            note       => 'test note',
+            group      => undef,
+            non_public => undef,
+            state => 'unconfirmed',
+        } } );
+
+        my $contact = $body->contacts->find({ category => 'grouped category' });
+        is $contact->get_extra_metadata('group'), undef, "group unset correctly";
+    };
+
 };